topk_ratio, normalized gradients, signed maxnorm
topk_ratio choice becomes more flexible. To decide the number of samples selected in topk algorithms, instead of ratio, we can just enter the number like 2. When 1 is entered, it is recognized as ratio and entire mini-batch is selected.
When --normalized_grads is entered, gradients are normalized in for correlation related algorithms. 
signed-norm algorithms are added. For example, in smaxnorm_topk, before selecting the topk, the sign of the correlation between the sample gradient and mini-batch gradient is calculated and multiplied by sample gradient's norm.